Scanlator can/cannot, adminAllScans and allScans

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->integer('role')->default(1);
+            $table->tinyInteger('role')->default(1);
             $table->integer('scanlator')->nullable()->default(null);
             $table->string('name');
             $table->string('email')->unique();
